requirements.txt
discord.py
python-dotenv
BeautifulSoup4
aiohttp